Add kubernetes support for Cassandra and Postgres

+++ core/src/main/java/pl/zabicki/billing/core/generator/ClientRequest.java
@@ -3,5 +3,5 @@ package pl.zabicki.billing.core.generator;
 /**
  * Request generation of n clients, each client has x accounts and each account has z events
  */
-public record ClientRequest(int numOfClients, int numOfAccounts, int numOfEventsPerAccount) {
+public record ClientRequest(long numOfClients, long numOfAccounts, long numOfEventsPerAccount) {
 }
